Create new files for optimized tasks query
I am creating new files for these changes rather than changing the existing
files because other branches are using the original files and errors would
result if these changes were made on the dev server.  Before this branch is
eventually merged, these changes should be made in the original files both in
repo and on the server, and these files should be deleted.