feat: implement disabled student filter.
This is a work in progress.  This file lives in the Slate repo and I am overridding it
here, but I am still unsure if this is a permanent solution, if the changes should be
moved to the Slate repo, or if I should create a custom section particiants request handler
for Slate-cbl.