fix: correct text clearing issue in task selector
the task selector used in the clone task and subtask fields of the task
form allow the user to type a string to filter results.  Currently,
when the user types a text string and results are returned, the query
text dissappears.  This creates confusion and does not allow the user
to continue typing to make their query string more precise.

This issues exists because the Task store does some post processing to
decorate results with subtask arrays and parent task references.  These
references are not necessary when populating the task selector, so I
altered the task selector to use the Task model rather than the task store,
bypassing this post-processing.

For specificity, the problem is caused by one specific line in the task store:
`task.set('SubTasks', subTasks, { dirty: false });`
the issue could also have been addressed by adding the silent flag to this
setter like this:
`task.set('SubTasks', subTasks, { dirty: false, silent: true });`
But the ExtJS docs caution against the use of this flag, it could theoretically
affect other parts of the app, and it seemed unnecessary in order to address
this specific issue.